refactor(challenge): Reduce Twig render time
Make ColumnsOfAnswer deferred.
