+++ templates/components/Challenge/ColumnsOfAnswer.html.twig
@@ -1,0 +1,5 @@
+<p {{ attributes }}>
+    <strong>輸出格式：</strong>欄位順序分別為：<code>{{
+        this.columnsOfAnswer|joinToQuoted('、')
+        }}</code>。
+</p>
